api/requirements/common.txt
Django==3.2.15

argon2-cffi==20.1.0
bigcommerce==0.22.3
boto3==1.17.95
django-anymail[mailgun]==8.4
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-graphql-extensions==0.1.0
django-graphql-jwt==0.3.2
django-hosts==4.0
django-import-export==2.5.0
django-ipware==3.0.2
django-localflavor==3.1
django-mathfilters==1.0.0
django-model-utils==4.1.1
django-ordered-model==3.4.3
django-redis==5.0.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
drf-nested-routers==0.93.3
graphene==3.0b7
graphene-django==3.0.0b7
graphql-core==3.1.5
graphql-relay==3.1.0
iso18245==1.1.0
mailchimp3==3.0.14
phonenumberslite==8.12.25
Pillow==8.2.0
psycopg2-binary==2.9
py2neo==2021.1.5
PyJWT==2.1.0
pynamodb==5.0.3
python-dateutil==2.8.1
python-tutuka-client==1.2.6
requests==2.25.1
requests-auth-aws-sigv4==0.7
djangorestframework-csv==2.1.1
ShopifyAPI==10.0.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.7.0
factory-boy==3.2.0
django-elasticsearch-dsl>=7.2.2

-r gateways.txt
-r private.txt
